[CCUBE-2183][GZ] differentiate filter addons types and internal types
