+++ src/filter/addons/filter-item-checkbox-helpers.ts
@@ -1,7 +1,5 @@
-import type {
-    FilterItemCheckboxOptionProps,
-    FlattenedFilterOption,
-} from "./types";
+import type { FlattenedFilterOption } from "./internal-types";
+import type { FilterItemCheckboxOptionProps } from "./types";
 
 // =============================================================================
 // HELPER FUNCTIONS FOR NESTED OPTIONS
