+++ src/filter/index.ts
@@ -1,6 +1,3 @@
-export type {
-    FilterItemCheckboxOptionProps,
-    FilterItemCheckboxProps,
-} from "./addons/types";
+export * from "./addons/types";
 export * from "./filter";
 export * from "./types";
